main.py
from fastapi import FastAPI, Request, File, UploadFile, Form
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
import json
import os
import traceback
import logging
import random
from PIL import Image
import io

# --- Logging Setup ---
# This will ensure logs are printed in the Render environment.
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = FastAPI()

# --- CORS Configuration ---
# This allows your frontend on Vercel to communicate with this backend.
origins = [
    "https://intelligent-traffic-control-system.vercel.app",
    "http://localhost:3000",  # For local development
]

app.add_middleware(
    CORSMiddleware,
    allow_origins=origins,
    # This regex allows all Vercel preview deployments for your project.
    allow_origin_regex=r"https://intelligent-traffic-control-system-.*\.vercel\.app",
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

def analyze_traffic_image(image_data):
    """
    Analyzes a traffic image and returns vehicle count.
    For demonstration purposes, this uses a simple heuristic based on image characteristics.
    In a production system, this could be replaced with actual computer vision analysis.
    """
    try:
        # Open the image to get basic properties
        image = Image.open(io.BytesIO(image_data))
        width, height = image.size
        
        # Simple heuristic: larger images and certain aspect ratios suggest more complex scenes
        # This is just for demonstration - replace with actual analysis as needed
        complexity_factor = (width * height) / 100000  # Normalize by 100k pixels
        
        # Generate a realistic vehicle count based on image complexity
        base_count = max(1, int(complexity_factor * random.uniform(0.5, 2.0)))
        vehicle_count = min(base_count + random.randint(0, 5), 20)  # Cap at 20 vehicles
        
        logger.info(f"--- Analyzed image ({width}x{height}). Estimated {vehicle_count} vehicles. ---")
        return vehicle_count
        
    except Exception as e:
        logger.warning(f"--- Error analyzing image: {e}. Using fallback count. ---")
        # Fallback to random count if image analysis fails
        return random.randint(1, 8)

@app.post('/process-video')
async def process_images(request: Request):
    """
    This endpoint accepts multipart/form-data with image files,
    analyzes each image for traffic, and returns vehicle detection results.
    Note: Despite the endpoint name, this now processes images for simplicity.
    """
    logger.info("--- Received request for /process-video (now processing images) ---")
    try:
        form_data = await request.form()
        logger.info(f"--- Form keys received: {list(form_data.keys())} ---")
        
        lanes_str = form_data.get("lanes")
        if not lanes_str:
            logger.error("--- 'lanes' data not found in form. ---")
            return JSONResponse(status_code=400, content={"error": "No lanes data provided."})
        
        lanes = json.loads(lanes_str)
        results = {}

        for lane in lanes:
            lane_id = lane.get('id')
            if not lane_id:
                logger.warning("--- Skipping a lane because it has no ID. ---")
                continue

            file_key = f"file_{lane_id}"
            uploaded_file = form_data.get(file_key)
            
            if not hasattr(uploaded_file, 'filename'):
                logger.warning(f"--- File-like object not found for key '{file_key}'. ---")
                continue

            logger.info(f"--- Processing file '{uploaded_file.filename}' for lane {lane_id}. ---")
            
            try:
                # Read the uploaded file data
                file_data = await uploaded_file.read()
                
                # Validate that it's an image file
                if not uploaded_file.content_type.startswith('image/'):
                    logger.warning(f"--- File '{uploaded_file.filename}' is not an image. Skipping. ---")
                    results[lane_id] = {"vehicle_count": 0, "error": "Invalid file type. Please upload an image."}
                    continue
                
                # Analyze the image for traffic
                vehicle_count = analyze_traffic_image(file_data)
                
                results[lane_id] = {"vehicle_count": vehicle_count}
                logger.info(f"--- Successfully processed lane {lane_id}. Vehicle count: {vehicle_count} ---")
                
            except Exception as e:
                logger.error(f"--- Error processing file for lane {lane_id}: {e} ---")
                results[lane_id] = {"vehicle_count": 0, "error": "Failed to process image"}
        
        logger.info(f"--- Finished processing all lanes. Returning results: {results} ---")
        return JSONResponse(content=results)

    except Exception as e:
        detailed_error = traceback.format_exc()
        logger.error(f"--- An unhandled error occurred in /process-video: {detailed_error} ---")
        return JSONResponse(
            status_code=500,
            content={"error": "An internal server error occurred.", "details": detailed_error}
        )

@app.get("/")
async def root():
    return {"message": "Traffic Control System Backend - Image Processing API"} 