requirements.txt
fastapi
uvicorn
python-multipart
numpy
requests
Pillow 